python/requirements.txt
asyncio
aiohttp
pipecat
pipecat-ai
python-dotenv
pydantic
daily